Add tests for party analysis fflogs processing also generalize fixtures in conftest
